[BE] chore: 유효성 검증, 예외 처리, 장소에 비용 추가 시 오류 수정 (#102)
* feat: 일정 등록, 수정, 삭제 기능 추가

* feat: 로그인 한 회원 조회 기능 추가

* feat: 각각 날짜에 대한 장소 등록, 수정, 삭제 기능 추가

* feat: 일정과 회원, 장소 연결 및 특정 일정 조회 기능 추가

* feat: 해당 유저가 작성한 전체 일정 조회

* fix: planMapper 병합 문제 수정

* fix: plan 응답 데이터 수정

* feat: city 조회 기능 추가

* fix: plan post-patch 응답 데이터 수정

* fix: plan patch 수정 오류 수정

* feat: planDates 응답 데이터 추가

* feat: ExceptionCode 응답 코드 추가

* feat: Redis 설정 추가

* feat: 장소-비용 연결

* feat: 예산-일정 연결

* fix: 일정 응답 데이터 수정(시간)

* feat: 배포자동화 관련 설정 파일 추가, 수정

* fix: 배포자동화 관련 설정 파일 수정

* fix: 응답 데이터 수정(날짜)

* fix: 장소-비용 등록, 수정 기능 수정

* feat: 장소-비용 삭제 기능 추가 및 응답 데이터 추가

* fix: 일정-예산 관계 설정 수정

* fix: 장소-비용 기능 수정

* fix: 세부 일정 조회 응답 데이터 수정

* fix: @SpringBootTest 주석 처리

* fix: 배포 자동화 설정 수정

* fix: 세부 일정 조회 시 권한 수정

* fix: 배포 자동화 설정 수정

* fix: 배포 자동화 설정 수정

* feat: Plan-City 연결 및 image 응답 추가

* feat: 장소 테이블 위도, 경도, 주소, 이미지 추가

* fix: 시간 응답 데이터 수정

* fix: 설정 파일 수정

* fix: 응답 데이터 수정

* Update gradle-server.yml

* Update gradle-server.yml

* fix: bootstrap.yml 삭제

* fix: 배포 설정 수정

* Update gradle-server.yml

* fix: place 이미지 주소 삭제 및 타입 변경

* fix: db 설정 수정

* feat: place 별점, 웹사이트, 번호, 영업시간 추가

* fix: place 응답 데이터 수정

* fix: Update gradle-server.yml

* fix: plan 응답 데이터에 city 이미지주소 추가

* feat: plan 시간 유효성 검증 추가

* chore: 예외처리 추가

* fix: 장소에 비용 추가 시 오류 수정

+++ server/src/main/java/com/newyear/mainproject/plan/service/PlanService.java
@@ -1,5 +1,6 @@
 package com.newyear.mainproject.plan.service;
 
+import com.newyear.mainproject.board.service.BoardService;
 import com.newyear.mainproject.city.CityService;
 import com.newyear.mainproject.exception.BusinessLogicException;
 import com.newyear.mainproject.exception.ExceptionCode;
@@ -10,6 +11,7 @@ import com.newyear.mainproject.plan.entity.PlanDates;
 import com.newyear.mainproject.plan.repository.PlanDateRepository;
 import com.newyear.mainproject.plan.repository.PlanRepository;
 import com.newyear.mainproject.util.DateCalculation;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
@@ -22,12 +24,14 @@ public class PlanService {
    private final PlanDateRepository planDateRepository;
    private final MemberService memberService;
    private final CityService cityService;
+   private final BoardService boardService;
 
-    public PlanService(PlanRepository planRepository, PlanDateRepository planDateRepository, MemberService memberService, CityService cityService) {
+    public PlanService(PlanRepository planRepository, PlanDateRepository planDateRepository, MemberService memberService, CityService cityService, @Lazy BoardService boardService) {
         this.planRepository = planRepository;
         this.planDateRepository = planDateRepository;
         this.memberService = memberService;
         this.cityService = cityService;
+        this.boardService = boardService;
     }
 
     /**
@@ -44,6 +48,11 @@ public class PlanService {
      */
     public Plan updatePlan(Plan plan) {
         Plan findPlan = findVerifiedPlan(plan.getPlanId());
+
+        //작성자만 수정 가능
+        if (!findPlan.getMember().equals(memberService.getLoginMember())) {
+            throw new BusinessLogicException(ExceptionCode.ACCESS_FORBIDDEN);
+        }
 
         Optional.ofNullable(plan.getPlanTitle())
                 .ifPresent(planTitle -> findPlan.setPlanTitle(planTitle));
@@ -103,6 +112,16 @@ public class PlanService {
      */
     public void deletePlan(Long planId) {
         Plan findPlan = findVerifiedPlan(planId);
+        //작성자만 삭제 가능
+        if (!findPlan.getMember().equals(memberService.getLoginMember())) {
+            throw new BusinessLogicException(ExceptionCode.ACCESS_FORBIDDEN);
+        }
+
+        //이 일정에 관련된 게시물이 있으면 게시물 삭제 먼저 하도록 예외 처리
+        if(!boardService.findPlanBoards(findPlan.getPlanId()).isEmpty()) {
+            throw new BusinessLogicException(ExceptionCode.BOARD_CHECK_EXISTS);
+        }
+
         planRepository.delete(findPlan);
     }
